Sync remaining upstream improvements
- deployer/tasks/setup.php: local setup support, dump file import,
  chooseSourceFile(), askConfirmation default yes, debug mode for local,
  auto-create DB on local, developer sync settings for local, mkdir -p
  for media dir, download path {{current_path}}, yrewrite uses
getMessage()
  Fork additions preserved: -c mysqldump flag, sandbox detection
- deployer/tasks/build/assets.php: replace simple yarn/gulp booleans
with
  upstream auto-detection (package.json, yarn/npm, webpack/gulp) and
  node_modules caching between builds

+++ deployer/tasks/setup.php
@@ -7,27 +7,41 @@ use Deployer\Host\Host;
 use Deployer\Host\Localhost;
 use Deployer\Task\Context;
 use Symfony\Component\Yaml\Yaml;
+
 use function count;
+use function is_string;
 use function YDeploy\downloadContent;
 use function YDeploy\onHost;
 use function YDeploy\uploadContent;
 
 desc('Setup redaxo instance');
 task('setup', new class() {
-    private $mysqlOptions;
-
     /** @var Host */
+    private $target;
+
+    /** @var bool */
+    private $local;
+
+    /** @var Host|string */
     private $source;
+
+    /** @var string */
+    private $mysqlOptions;
 
     /** @var null|string */
     private $server;
 
     public function __invoke(): void
     {
-        cd('{{release_path}}');
+        $this->target = Context::get()->getHost();
+        $this->local = $this->target instanceof Localhost;
 
-        if (test('[ -f {{data_dir}}/core/config.yml ]')) {
-            return;
+        if (!$this->local) {
+            cd('{{release_path}}');
+
+            if (test('[ -f {{data_dir}}/core/config.yml ]')) {
+                return;
+            }
         }
 
         writeln('');
@@ -47,25 +61,41 @@ task('setup', new class() {
         }
     }
 
-    private function chooseSource(): Host
+    private function chooseSource()
     {
         $this->headline('Setup <fg=cyan>{{hostname}}</fg=cyan>');
 
         $hosts = Deployer::get()->hosts;
-        $localhost = new Localhost('local');
 
-        if (count($hosts) < 2) {
-            writeln("The host <fg=cyan>{{hostname}}</fg=cyan> will be initialized by data from <fg=cyan>{$localhost}</fg=cyan>.");
-            writeln('');
+        if ($this->local && count($hosts) < 2) {
+            writeln('The data must be imported from an existing dump file.');
 
-            return $localhost;
+            return $this->chooseSourceFile();
         }
 
-        $hosts->set($localhost->getHostname(), $localhost);
+        if ($this->local) {
+            writeln('The data can be imported from one of the hosts, or from dump file.');
+            writeln('');
+
+            if (!askConfirmation('Import from host', true)) {
+                return $this->chooseSourceFile();
+            }
+
+            writeln('');
+        }
 
         $hostsArray = $hosts->all();
-        unset($hostsArray[Context::get()->getHost()->getHostname()]);
+        unset($hostsArray[$this->target->getAlias()]);
         $hostsArray = array_keys($hostsArray);
+
+        if (count($hostsArray) < 2) {
+            $host = $hosts->get($hostsArray[0]);
+
+            writeln("The host <fg=cyan>{{hostname}}</fg=cyan> will be initialized by data from <fg=cyan>{$host}</fg=cyan>.");
+            writeln('');
+
+            return $host;
+        }
 
         writeln('The data from which host shall be used for initializing <fg=cyan>{{hostname}}</fg=cyan>?');
         writeln('');
@@ -77,12 +107,31 @@ task('setup', new class() {
         return $hosts->get($host);
     }
 
+    private function chooseSourceFile(): string
+    {
+        writeln('');
+
+        while (true) {
+            $file = ask('Dump file', getcwd().'/dump.sql');
+            writeln('');
+
+            if (file_exists($file)) {
+                return $file;
+            }
+
+            writeln('<error>The file does not exist: '.$file.'</error>');
+            writeln('');
+        }
+    }
+
     private function setConfigYml(): void
     {
         $this->headline('Create config.yml for <fg=cyan>{{hostname}}</fg=cyan>');
 
         if ($this->source instanceof Localhost) {
             $config = file_get_contents(getcwd().'/'.get('data_dir').'/core/config.yml');
+        } elseif (is_string($this->source)) {
+            $config = file_get_contents(getcwd().'/'.get('src_dir').'/core/default.config.yml');
         } else {
             $config = onHost($this->source, static function () {
                 return downloadContent('{{data_dir}}/core/config.yml');
@@ -92,7 +141,7 @@ task('setup', new class() {
         $config = Yaml::parse($config);
 
         $config['setup'] = false;
-        $config['debug'] = false;
+        $config['debug'] = $this->local;
         $config['instname'] = 'rex'.date('YmdHis');
 
         $config['server'] = ask('Server:', get('url'));
@@ -121,18 +170,28 @@ task('setup', new class() {
                 '--host='.escapeshellarg($db['host']),
                 '--user='.escapeshellarg($db['login']),
                 '--password='.escapeshellarg($db['password']),
-                escapeshellarg($db['name']),
+                $this->local ? '' : escapeshellarg($db['name']),
             ]));
 
             try {
-                run('< '.escapeshellarg($this->mysqlOptions).' xargs {{bin/mysql}} -e ";"');
+                $cmd = $this->local ? 'CREATE DATABASE IF NOT EXISTS '.escapeshellcmd($db['name']).' CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci' : '';
+                run('< '.escapeshellarg($this->mysqlOptions).' xargs {{bin/mysql}} -e "'.$cmd.';"');
                 $dbValid = true;
             } catch (RunException $e) {
                 writeln('');
-                writeln('<error>Could not connect to database: '.trim($e->getErrorOutput()).'</error>');
+                writeln('<error>Could not '.($this->local ? 'create/' : '').'connect to database: '.trim($e->getErrorOutput()).'</error>');
                 writeln('');
             }
         } while (!$dbValid);
+
+        if ($this->local) {
+            file_put_contents($this->mysqlOptions, implode("\n", [
+                '--host='.escapeshellarg($db['host']),
+                '--user='.escapeshellarg($db['login']),
+                '--password='.escapeshellarg($db['password']),
+                escapeshellarg($db['name']),
+            ]));
+        }
 
         $config['db'][1] = $db;
 
@@ -147,46 +206,63 @@ task('setup', new class() {
 
     private function copyDatabase(): void
     {
-        $this->headline("Copy database from <fg=cyan>{$this->source}</fg=cyan> to <fg=cyan>{{hostname}}</fg=cyan>");
+        if (is_string($this->source)) {
+            $this->headline('Import database dump');
 
-        $path = get('data_dir').'/addons/ydeploy/'.date('YmdHis').'.sql';
+            $path = $this->source;
+        } else {
+            $this->headline("Copy database from <fg=cyan>{$this->source}</fg=cyan> to <fg=cyan>{{hostname}}</fg=cyan>");
 
-        if (!$this->source instanceof Localhost) {
-            cd('{{current_path}}');
+            $path = get('data_dir').'/addons/ydeploy/'.date('YmdHis').'.sql';
+
+            // export source database
+            onHost($this->source, function () use ($path) {
+                if (!$this->source instanceof Localhost) {
+                    cd('{{current_path}}');
+                }
+
+                $sqlFile = escapeshellarg($path);
+                run('{{bin/php}} {{bin/console}} db:connection-options | xargs {{bin/mysqldump}} -c > '.$sqlFile);
+
+                if (str_contains(run('head -n 1 '.$sqlFile), 'sandbox')) {
+                    info('Detected MariaDB sandbox mode: Removing');
+                    run("sed -i '1d' ".$sqlFile);
+                }
+
+                if ($this->source instanceof Localhost) {
+                    return;
+                }
+
+                download("{{current_path}}/$path", $path);
+                run('rm -f '.escapeshellarg($path));
+            });
         }
 
-        // export source database
-        onHost($this->source, static function () use ($path) {
-            $sqlFile = escapeshellarg($path);
-            run('{{bin/php}} {{bin/console}} db:connection-options | xargs {{bin/mysqldump}} -c > ' . $sqlFile);
-            if (str_contains(run('head -n 1 ' . $sqlFile), 'sandbox')) {
-                info('Detected MariaDB sandbox mode: Removing');
-                run("sed -i '1d' " . $sqlFile);
-            }
-
-            if (Context::get()->getHost() instanceof Localhost) {
-                return;
-            }
-
-            download("{{release_path}}/$path", $path);
-            run('rm -f '.$sqlFile);
-        });
-
         // upload and import the dump
-        upload($path, "{{release_path}}/$path");
+        if (!$this->local) {
+            upload($path, "{{release_path}}/$path");
+        }
         run('< '.escapeshellarg($this->mysqlOptions).' xargs sh -c \'{{bin/mysql}} "$0" "$@" < '.escapeshellcmd(escapeshellarg($path)).'\'');
 
-        run('rm -f '.escapeshellarg($path));
-        unlink(getcwd().'/'.$path);
+        if (!$this->local) {
+            run('rm -f '.escapeshellarg($path));
+        }
+        if (!is_string($this->source)) {
+            unlink(getcwd().'/'.$path);
+        }
 
         $this->ok();
     }
 
     private function configureDeveloper(): void
     {
-        $this->headline('Configure developer addon for production usage');
+        $this->headline('Configure developer addon for '.($this->local ? 'local' : 'production').' usage');
 
-        run('< '.escapeshellarg($this->mysqlOptions).' xargs {{bin/mysql}} -e "UPDATE rex_config SET value = \"false\" WHERE namespace=\"developer\" AND \`key\` NOT IN (\"templates\", \"modules\", \"actions\", \"items\")"');
+        if ($this->local) {
+            run('< '.escapeshellarg($this->mysqlOptions).' xargs {{bin/mysql}} -e "UPDATE rex_config SET value = \"true\" WHERE namespace=\"developer\" AND \`key\` IN (\"sync_frontend\", \"sync_backend\", \"rename\", \"dir_suffix\", \"delete\")"');
+        } else {
+            run('< '.escapeshellarg($this->mysqlOptions).' xargs {{bin/mysql}} -e "UPDATE rex_config SET value = \"false\" WHERE namespace=\"developer\" AND \`key\` NOT IN (\"templates\", \"modules\", \"actions\", \"items\")"');
+        }
 
         $this->ok();
     }
@@ -197,7 +273,7 @@ task('setup', new class() {
             $data = run('< '.escapeshellarg($this->mysqlOptions).' xargs {{bin/mysql}} --silent --raw --skip-column-names -e "SELECT id, domain FROM rex_yrewrite_domain"');
             $data = trim($data);
         } catch (RunException $exception) {
-            if (str_contains($exception->getErrorOutput(), 'ERROR 1146')) {
+            if (str_contains($exception->getMessage(), 'ERROR 1146')) {
                 // Table does not exist (yrewite not activated)
                 return;
             }
@@ -224,6 +300,10 @@ task('setup', new class() {
 
     private function copyMedia(): void
     {
+        if (is_string($this->source)) {
+            return;
+        }
+
         $this->headline("Copy media files from <fg=cyan>{$this->source}</fg=cyan> to <fg=cyan>{{hostname}}</fg=cyan>");
 
         $path = get('data_dir').'/addons/ydeploy/media_'.date('YmdHis').'.tar.gz';
@@ -237,7 +317,7 @@ task('setup', new class() {
             }
 
             try {
-                download("{{release_path}}/$path", $path);
+                download("{{current_path}}/$path", $path);
             } finally {
                 run('rm -f '.escapeshellarg($path));
             }
@@ -245,6 +325,7 @@ task('setup', new class() {
 
         try {
             upload($path, "{{release_path}}/$path");
+            run('mkdir -p {{media_dir}}');
             run('tar -zxvf '.escapeshellarg($path).' -C {{media_dir}}/');
         } finally {
             unlink($path);
